--- src/ui/course-card/homeCardItems.ts
@@ -1,25 +1,0 @@
-import { CardItems } from "../../interfaces/card-items/CardItems";
-
-export const homeCardItems: CardItems[] = [
-  {
-    id: 1,
-    src: "https://avatars.dzeninfra.ru/get-zen_brief/271828/pub_659979451225e85516f36401_659979461225e85516f36402/scale_2400",
-    text: "In this course, you will understand how HTML works and write your own website. The best course for those who want to understand the process of creating websites.",
-    lang: "https://cdn-icons-png.flaticon.com/128/330/330437.png",
-    href: "https://youtu.be/W4MIiV4nZDY?si=-nVtg6Gj1bMMrQns",
-  },
-  {
-    id: 2,
-    src: "https://avatars.dzeninfra.ru/get-zen_brief/271828/pub_659979451225e85516f36401_659979461225e85516f36402/scale_2400",
-    text: "The best beginner's course on JS, from one of the best authors on the subject.",
-    lang: "https://cdn-icons-png.flaticon.com/128/330/330437.png",
-    href: "https://youtu.be/CxgOKJh4zWE?si=d41vrgPCFus4fD7z",
-  },
-  {
-    id: 3,
-    src: "https://avatars.dzeninfra.ru/get-zen_brief/271828/pub_659979451225e85516f36401_659979461225e85516f36402/scale_2400",
-    text: "The best beginner's course on JS, from one of the best authors on the subject.",
-    lang: "https://cdn-icons-png.flaticon.com/128/330/330437.png",
-    href: "https://youtu.be/CxgOKJh4zWE?si=d41vrgPCFus4fD7z",
-  },
-];
